Make the lint and type-check CI gates pass deterministically
The Lint and Type Check jobs were red on every push (and had been red on
main before this work), for environment reasons the local box masked:

Lint: with `extend-select`, a newer ruff (CI installs 0.16.x) pulls its
larger default rule set into scope and flagged ~970 findings, while the
pinned 0.15.x used locally did not. Switched to an explicit `select`
(E4/E7/E9/F + B + S108/S110) so the rule set is fixed regardless of ruff
version.

Type Check: mypy pinned to `python_version = "3.10"` rejected the PEP 695
`type` statement in newer numpy stubs as 3.12+ syntax, aborting the check.
Removed the pin (mypy uses the running interpreter; the test matrix and
ruff's py310 target still guard 3.10 compatibility). numpy's stubs also
tightened across 2.2/2.4/2.5, surfacing float64/bool/ndarray typing issues
that only appeared under some versions; fixed them at the source with
`float()`/`bool()`/`int()` casts and small annotations in both packages
(hackrf TX buffer, the noise-reduction overlap-add, the CW gap calc, the
format-detector score, and the antenna calibration/cross-correlator/doa
call sites), added `# type: ignore` to the antenna optional-import
fallbacks, and set `warn_unused_ignores = false` so a `type: ignore` needed
under one numpy stub is not flagged as unused under another.

Verified clean under numpy 2.2.6, 2.4.6 and 2.5.2 with mypy 1.19.1 and 2.3.1,
and under ruff 0.15.8 and 0.16.6. 1011 tests pass; build + twine --strict
still pass.

+++ src/sdr_module/dsp/demodulators.py
@@ -1544,13 +1544,13 @@ class CWDemodulator(Demodulator):
         for i, (start, end) in enumerate(zip(key_down, key_up, strict=True)):
             # Check gap before this element
             if i == 0 and self._last_key_time > 0:
-                gap = (
+                gap = float(
                     start / self._sample_rate
                     + (current_time - len(keying) / self._sample_rate)
                     - self._last_key_time
                 )
             elif i > 0:
-                gap = (start - key_up[i - 1]) / self._sample_rate
+                gap = float((start - key_up[i - 1]) / self._sample_rate)
             else:
                 gap = 0.0
